requirements.txt
appdirs
pygetwindow
obs-websocket-py
git+https://github.com/pybluez/pybluez
